hf-predictor/app/requirements.txt
requests
fastapi
uvicorn[standard]
redis
pyyaml
